pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "google-workspace-mcp-advanced"
version = "1.0.0"
description = "Advanced Google Workspace MCP Server - Gmail, Calendar, Drive, Docs, Sheets, Slides, Forms, Tasks, Chat, Search with bidirectional sync"
readme = "README.md"
keywords = ["mcp", "google", "workspace", "llm", "ai", "claude", "model", "context", "protocol", "server"]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Office/Business",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed"
]
dependencies = [
    "fastapi>=0.115.12",
    "fastmcp>=2.14.0",
    "google-api-python-client>=2.168.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "httpx>=0.28.1",
    "markdown-it-py>=3.0.0",  # CommonMark parser for Markdown-to-Docs conversion
    "mdit-py-plugins>=0.5.0",  # Extensions for markdown-it-py (tasklists, etc.)
    "py-key-value-aio>=0.3.0",
    "pyasn1>=0.6.2",  # CVE-2026-23490: DoS fix
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "cryptography>=45.0.0",
]

[[project.authors]]
name = "Google Workspace MCP Contributors"
email = "contributions@example.com"

[project.license]
text = "MIT"

[project.urls]
Repository = "https://github.com/Skeptomenos/google-workspace-mcp-advanced"

[project.scripts]
google-workspace-mcp-advanced = "main:main"

[project.optional-dependencies]
valkey = [
    "py-key-value-aio[valkey]>=0.3.0",
]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "requests>=2.32.3",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "requests>=2.32.3",
    "pyright>=1.1.408",
    "ruff>=0.12.4",
    "pyright>=1.1.408",
]

[tool.setuptools]
py-modules = ["fastmcp_server", "main"]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*", "docs*", "build", "dist", "old*", "specs*"]

[tool.setuptools.package-data]
core = ["tool_tiers.yaml"]
"*" = ["py.typed"]

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101"]  # Skip assert warnings (used in tests)

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501", "B008"]  # B008: Allow Body()/Field() in defaults (FastAPI pattern)

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "mcp_protocol: MCP stdio protocol-level smoke tests",
    "live_mcp: Live Google Workspace integration tests (network and credentials required)",
    "live_write: Live mutation tests against controlled artifacts"
]
